pyproject.toml
[tool.poetry]
name = "linq_backend"
version = "0.1.0"
description = ""
authors = ["NathanDBrowne <nathan.browne1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.2"
requests = "^2.28.0"
loguru = "^0.6.0"
pandas = "^1.4.2"
pymongo = {extras = ["srv"], version = "^4.1.1"}
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
